add helper for telemetry
